+++ presenter/src/main/java/com/mashup/twotoo/presenter/designsystem/component/bottomsheet/SetImageOptionDialog.kt
@@ -22,6 +22,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Dialog
 import androidx.compose.ui.window.DialogProperties
+import com.mashup.twotoo.presenter.designsystem.component.common.TextContainer
 import com.mashup.twotoo.presenter.designsystem.theme.TwoTooTheme
 
 /**
@@ -106,19 +107,6 @@ fun SetImageOptionDialog(
                 )
             }
         }
-    }
-}
-
-@Composable
-fun TextContainer(
-    text: @Composable () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    Box(
-        modifier = modifier,
-        contentAlignment = Alignment.Center,
-    ) {
-        text()
     }
 }
