+++ presenter/src/main/java/com/mashup/twotoo/presenter/designsystem/component/common/Containers.kt
@@ -1,0 +1,19 @@
+package com.mashup.twotoo.presenter.designsystem.component.common
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+
+@Composable
+fun TextContainer(
+    text: @Composable () -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    Box(
+        modifier = modifier,
+        contentAlignment = Alignment.Center,
+    ) {
+        text()
+    }
+}
